MAHOUT-1570: Flink: added headers, comments and aknowledgements

+++ flink/src/main/scala/org/apache/mahout/flinkbindings/package.scala
@@ -1,3 +1,21 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package org.apache.mahout
 
 import scala.reflect.ClassTag
@@ -57,24 +75,25 @@ package object flinkbindings {
   private[flinkbindings] implicit def unwrapFromWritable(w: MatrixWritable): Matrix = w.get()
   private[flinkbindings] implicit def unwrapFromWritable(w: VectorWritable): Vector = w.get()
 
+
   def readCsv(file: String, delim: String = ",", comment: String = "#")
-             (implicit dc: DistributedContext): CheckpointedDrm[Long] = {
+             (implicit dc: DistributedContext): CheckpointedDrm[Int] = {
     val vectors = dc.env.readTextFile(file)
-        .filter(new FilterFunction[String] {
-          def filter(in: String): Boolean = {
-            !in.startsWith(comment)
-          }
-        })
-        .map(new MapFunction[String, Vector] {
-          def map(in: String): Vector = {
-            val array = in.split(delim).map(_.toDouble)
-            new DenseVector(array)
-          }
-        })
+      .filter(new FilterFunction[String] {
+        def filter(in: String): Boolean = {
+          !in.startsWith(comment)
+        }
+      })
+      .map(new MapFunction[String, Vector] {
+        def map(in: String): Vector = {
+          val array = in.split(delim).map(_.toDouble)
+          new DenseVector(array)
+        }
+      })
     datasetToDrm(vectors)
   }
 
-  def datasetToDrm(ds: DataSet[Vector]): CheckpointedDrm[Long] = {
+  def datasetToDrm(ds: DataSet[Vector]): CheckpointedDrm[Int] = {
     val zipped = new DataSetOps(ds).zipWithIndex
     datasetWrap(zipped)
   }
